Issue-2121 add an 'enable' option in preHandler
